+++ src/hash/mod.rs
@@ -1,3 +1,4 @@
+pub mod ahash;
 pub mod algorithms;
 pub mod dhash;
 pub mod phash;